Renamed core sdk into norman api.
Resorted imports after core and util library splits.

(#86c778p0x)
