+++ norman_api/clients/http_client.py
@@ -4,12 +4,13 @@ from typing import Optional, Type
 
 import httpx
 from httpx import Response
-from norman_core._app_config import AppConfig
-from norman_core.clients.objects.request_kwargs import RequestKwargs
-from norman_core.clients.objects.response_encoding import ResponseEncoding
 from norman_objects.shared.security.sensitive import Sensitive
 from norman_utils.singleton import Singleton
 from typing_extensions import Unpack
+
+from norman_api._app_config import AppConfig
+from norman_api.clients.objects.request_kwargs import RequestKwargs
+from norman_api.clients.objects.response_encoding import ResponseEncoding
 
 
 class HttpClient(metaclass=Singleton):
