+++ norman_api/clients/objects/request_encoding.py
