+++ norman_api/clients/objects/response_encoding.py
