+++ norman_api/services/authenticate/login.py
@@ -1,10 +1,11 @@
-from norman_core.clients.http_client import HttpClient
 from norman_objects.services.authenticate.login.account_id_password_login_request import AccountIDPasswordLoginRequest
 from norman_objects.services.authenticate.login.api_key_login_request import ApiKeyLoginRequest
 from norman_objects.services.authenticate.login.email_password_login_request import EmailPasswordLoginRequest
 from norman_objects.services.authenticate.login.login_response import LoginResponse
 from norman_objects.services.authenticate.login.name_password_login_request import NamePasswordLoginRequest
 from norman_utils.singleton import Singleton
+
+from norman_api.clients.http_client import HttpClient
 
 
 class Login(metaclass=Singleton):
