+++ norman_api/services/authenticate/logout.py
@@ -1,6 +1,7 @@
-from norman_core.clients.http_client import HttpClient
 from norman_objects.shared.security.sensitive import Sensitive
 from norman_utils.singleton import Singleton
+
+from norman_api.clients.http_client import HttpClient
 
 
 class Logout(metaclass=Singleton):
