+++ norman_api/services/authenticate/signup.py
@@ -1,4 +1,3 @@
-from norman_core.clients.http_client import HttpClient
 from norman_objects.services.authenticate.login.login_response import LoginResponse
 from norman_objects.services.authenticate.signup.signup_email_request import SignupEmailRequest
 from norman_objects.services.authenticate.signup.signup_key_request import SignupKeyRequest
@@ -6,6 +5,8 @@ from norman_objects.services.authenticate.signup.signup_key_response import Sign
 from norman_objects.services.authenticate.signup.signup_password_request import SignupPasswordRequest
 from norman_objects.shared.accounts.account import Account
 from norman_utils.singleton import Singleton
+
+from norman_api.clients.http_client import HttpClient
 
 
 class Signup(metaclass=Singleton):
