+++ norman_api/services/file_pull/file_pull.py
@@ -1,4 +1,3 @@
-from norman_core.clients.http_client import HttpClient
 from norman_objects.services.file_pull.download.tracked_download_union import TrackedDownloadUnion
 from norman_objects.services.file_pull.requests.asset_download_request import AssetDownloadRequest
 from norman_objects.services.file_pull.requests.input_download_request import InputDownloadRequest
@@ -6,6 +5,8 @@ from norman_objects.services.file_pull.requests.output_download_request import O
 from norman_objects.shared.security.sensitive import Sensitive
 from norman_utils.singleton import Singleton
 from pydantic import TypeAdapter
+
+from norman_api.clients.http_client import HttpClient
 
 
 class FilePull(metaclass=Singleton):
