+++ norman_api/services/file_push/file_push.py
@@ -1,10 +1,11 @@
-from norman_core.clients.http_client import HttpClient
 from norman_objects.services.file_push.checksum.checksum_request import ChecksumRequest
 from norman_objects.services.file_push.pairing.socket_asset_pairing_request import SocketAssetPairingRequest
 from norman_objects.services.file_push.pairing.socket_input_pairing_request import SocketInputPairingRequest
 from norman_objects.services.file_push.pairing.socket_pairing_response import SocketPairingResponse
 from norman_objects.shared.security.sensitive import Sensitive
 from norman_utils.singleton import Singleton
+
+from norman_api.clients.http_client import HttpClient
 
 
 class FilePush(metaclass=Singleton):
