+++ norman_api/services/retrieve/retrieve.py
@@ -1,9 +1,10 @@
 from typing import AsyncIterator, Tuple
 
 import httpx
-from norman_core.clients.http_client import HttpClient, ResponseEncoding
 from norman_objects.shared.security.sensitive import Sensitive
 from norman_utils.singleton import Singleton
+
+from norman_api.clients.http_client import HttpClient, ResponseEncoding
 
 
 class Retrieve(metaclass=Singleton):
